+++ R/generics.R
@@ -1,35 +1,35 @@
 #' @param x An object of class `scdrake_list`.
-#' @param name,names_ A character scalar or vector: name / names of element(s) to extract.
+#' @param i A character scalar or vector: name / names of element(s) to extract.
 #' @param check A logical scalar: if `TRUE`, throw an error when non-existing names or subscripts are present.
 #' @name scdrake_list_generics_parameters
 NULL
 
 #' @rdname scdrake_list
 #' @export
-`$.scdrake_list` <- function(x, name) {
+`$.scdrake_list` <- function(x, i) {
   x_exprs <- deparse(substitute(x))
-  assert_that_(name %in% names(x), msg = "Variable {.field {name}} not found in {.code {x_exprs}}")
-  get(name, x)
+  assert_that_(i %in% names(x), msg = "Variable {.field {i}} not found in {.code {x_exprs}}")
+  get(i, x)
 }
 
 #' @rdname scdrake_list
 #' @export
-`[.scdrake_list` <- function(x, names_, check = TRUE) {
+`[.scdrake_list` <- function(x, i, check = TRUE) {
   x_exprs <- deparse(substitute(x))
   x <- unclass(x)
 
   if (check) {
-    if (is_character(names_)) {
-      names_exist <- names_ %in% names(x)
+    if (is_character(i)) {
+      names_exist <- i %in% names(x)
       assert_that_(
         all(names_exist),
         msg = str_space(
           "{if (sum(!names_exist) == 1) 'Variable' else 'Variables'}",
-          "{.field {names_[!names_exist]}} not found in {.code {x_exprs}}"
+          "{.field {i[!names_exist]}} not found in {.code {x_exprs}}"
         )
       )
-    } else if (is_bare_numeric(names_)) {
-      subscripts <- as.integer(names_)
+    } else if (is_bare_numeric(i)) {
+      subscripts <- as.integer(i)
       subscripts_exist <- subscripts %in% seq_along(x)
       assert_that_(
         all(subscripts_exist),
@@ -42,8 +42,8 @@ NULL
       )
     }
   } else {
-    if (is_character(names_)) {
-      for (name in names_) {
+    if (is_character(i)) {
+      for (name in i) {
         if (!name %in% names(x)) {
           x[name] <- list(NULL)
         }
@@ -51,26 +51,26 @@ NULL
     }
   }
 
-  structure(x[names_], class = "scdrake_list")
+  structure(x[i], class = c("scdrake_list", "list"))
 }
 
 #' @rdname scdrake_list
 #' @export
-`[[.scdrake_list` <- function(x, name, check = TRUE) {
+`[[.scdrake_list` <- function(x, i, check = TRUE) {
   x_exprs <- deparse(substitute(x))
   x <- unclass(x)
 
   if (check) {
-    if (is_bare_numeric(name)) {
+    if (is_bare_numeric(i)) {
       possible_subscripts <- seq_along(x)
-      assert_that_(name %in% possible_subscripts, msg = "Subscript {.field {name}} is out of bounds for {.code {x_exprs}}")
-      x <- x[[name]]
+      assert_that_(i %in% possible_subscripts, msg = "Subscript {.field {i}} is out of bounds for {.code {x_exprs}}")
+      x <- x[[i]]
     } else {
-      assert_that_(name %in% names(x), msg = "Variable {.field {name}} not found in {.code {x_exprs}}")
-      x <- get(name, x)
+      assert_that_(i %in% names(x), msg = "Variable {.field {i}} not found in {.code {x_exprs}}")
+      x <- get(i, x)
     }
   } else {
-    x <- x[[name]]
+    x <- x[[i]]
   }
 
   return(x)