+++ tests/testthat/test-generics.R
@@ -38,12 +38,11 @@ test_that("overloaded [ operator works for scdrake_list class", {
   expect_identical(l[c("a", "b")], scdrake_list(list(a = 1, b = 2)))
   expect_identical(l[c(1, 3)], scdrake_list(list(a = 1, c = NULL)))
 
-
   expect_error(l["d"])
   expect_identical(l["d", check = FALSE], scdrake_list(list(d = NULL)))
   expect_error(l[4])
   expect_error(l[c(1, 4)])
-  expect_identical(l[4, check = FALSE], structure(list(NULL), .Names = NA_character_, class = "scdrake_list"))
+  expect_identical(l[4, check = FALSE], structure(list(NULL), .Names = NA_character_, class = c("scdrake_list", "list")))
 
   l[c("z", "y")] <- c(1, 2)
   expect_s3_class(l, "scdrake_list")